refactor: 🔄 Rename following naming conventions(#29)
